[smduarte, merge-from-trunk]  Ported sys.* to use the standard logger (as used in swing.*).
  Logging levels are now settable through the corresponding .properties file.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/concordant/sources/swiftcloud/java/SwiftCloud-branches/1pc@1284 049c2a90-bdf1-4f83-8191-3719d539f8e0

+++ src/sys/utils/Threading.java
@@ -42,15 +42,15 @@ public class Threading {
 	}
 
 	
-	static public void sleep(int ms) {
+	static public void sleep(long ms) {
 		try {
 			Thread.sleep(ms);
 		} catch (InterruptedException x) {
 			x.printStackTrace();
 		}
 	}
-
-	static public void sleep(int ms, int ns) {
+	
+	static public void sleep(long ms, int ns) {
 		try {
 			Thread.sleep(ms, ns);
 		} catch (InterruptedException x) {
@@ -93,7 +93,7 @@ public class Threading {
 		}
 	}
 
-	static public void synchronizedWaitOn(Object o, int ms) {
+	static public void synchronizedWaitOn(Object o, long ms) {
 		synchronized (o) {
 			try {
 				o.wait(ms);
